Feat(index.html): Add extra classes to elements with common properties and include empty "span" element for desktop design
This commit adds extra classes to elements that share common properties in order to improve code organization and maintainability.
By applying these additional classes, it becomes easier to manage and update the styling of these elements in the future.
Furthermore, an empty "span" element has been added specifically for the desktop design.
This element will be utilized to achieve a desired layout or visual effect on desktop screens.