Docs(designs): Create 'designs' folder inside 'src' directory and add desktop and mobile designs
This commit moves the creation of the 'designs' folder from the project root directory to the 'src' directory.
Inside the 'src/designs' folder, two design files, one for desktop and one for mobile, have been added.
These files serve the purpose of showcasing the project's visual appearance in the project documentation.